+++ src/main/java/Main.java
@@ -1,19 +1,22 @@
 import UI.*;
+import UI.create_account_UI.CreateAccountUI;
+import UI.login_UI.LoginUI;
+import UI.task_display_UI.*;
 import config.CommonCryptograph;
 import config.Cryptograph;
 import config.UserDataStoreGateway;
 import config.UserStorage;
-import create_task_UI.CreateTaskUI;
-import dashboard_ui.DashboardUI;
+import UI.create_task_UI.CreateTaskUI;
+import UI.dashboard_UI.DashboardUI;
 import entity_factories.*;
 import entity_layer.PasswordSuggester;
 import entity_layer.RandomPasswordGenerator;
 import task_display_ui.*;
-import task_tree_UI.TaskTreeUI;
+import UI.task_tree_UI.TaskTreeUI;
 import use_cases.add_task.*;
 import use_cases.complete_task.*;
-import time_manager_UI.SetAvailabilityUI;
-import time_manager_UI.DisplayAvailabilityTimeBlockUI;
+import UI.time_manager_UI.SetAvailabilityUI;
+import UI.time_manager_UI.DisplayAvailabilityTimeBlockUI;
 import use_cases.create_curriculum.*;
 import use_cases.display_availability_timeblocks.DisplayAvailabilityTimeBlockController;
 import use_cases.display_availability_timeblocks.DisplayAvailabilityTimeBlockInteractor;
