+++ src/main/java/UI/create_task_UI/CreateTaskUI.java
@@ -1,4 +1,4 @@
-package create_task_UI;
+package UI.create_task_UI;
 
 import UI.*;
 import entity_factories.CommonTaskFactory;
