+++ src/main/java/UI/dashboard_UI/DashboardUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="dashboard_ui.DashboardUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="UI.dashboard_UI.DashboardUI">
   <grid id="27dc6" binding="Dashboard" layout-manager="GridLayoutManager" row-count="6" column-count="5" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
