+++ src/main/java/UI/task_display_UI/TaskUI.java
@@ -1,4 +1,4 @@
-package task_display_ui;
+package UI.task_display_UI;
 
 import UI.CentralWindow;
 import UI.WindowManager;
