+++ src/main/java/UI/task_display_UI/TaskUiInputBoundary.java
@@ -1,4 +1,4 @@
-package task_display_ui;
+package UI.task_display_UI;
 
 /**
  * An input boundary interface implemented by the TaskUiInteractor class.
