+++ src/main/java/UI/task_display_UI/TaskUiInteractor.java
@@ -1,4 +1,4 @@
-package task_display_ui;
+package UI.task_display_UI;
 
 import entity_layer.Curriculum;
 import entity_layer.InMemoryUser;
