+++ src/main/java/UI/task_display_UI/TaskUiModel.java
@@ -1,4 +1,4 @@
-package task_display_ui;
+package UI.task_display_UI;
 
 /**
  * Model class that packages the required information used by TaskUiModelPresenter
