+++ src/main/java/UI/time_manager_UI/DisplayAvailabilityTimeBlockUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="time_manager_UI.DisplayAvailabilityTimeBlockUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="UI.time_manager_UI.DisplayAvailabilityTimeBlockUI">
   <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="3" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
