+++ src/test/java/task_display_ui/TaskUiTestPresenter.java
@@ -1,5 +1,9 @@
 package task_display_ui;
 
+import UI.task_display_UI.TaskUiModel;
+import UI.task_display_UI.TaskUiOutputBoundary;
+import UI.task_display_UI.TaskUiViewInterface;
+
 public class TaskUiTestPresenter implements TaskUiOutputBoundary {
 
     private TaskUiModel taskUiModel;
